requirements.txt
cairosvg
pillow
openai
azure-identity
python-dotenv
